Add PerResourcePositionsService
